Add a much faster implementation of Day 18.
Like the previous solution, this processes all explosions in a single pass,
but this one also uses the fac that if a split happens at MAX_DEPTH, then the
newly created node would immediately explode on the next iteration, so
instead of creating it, I simulate the combined split + explode.

This allows processing all splits in a single pass too, resulting in an O(N)
reduce implementation.