Add more efficient solution to Day 17 (and rename the old one to 17-slower.py)
The efficient solution uses (row, col, last_dir) as the state. This is slightly
faster than the previous solution, which used (row, col, last_dir, repeat),
because it generates smaller graphs. Not only is the number of vertices reduced
by a factor 3 (part 1) or 10 (part 2), but we can also reduce the total number
of edges, because the old solution added some edges from equivalent states: if
we move from (row, col, dir, rep) to (row2, col2, dir2, 1) where dir ≠ dir2,
then the old value of `rep` does not matter, so in the old solution I would
generate 3 to 10 edges like this, versus only 1 in the new solution.