Take arguments by value instead of by const-pointer.
This lets the compiler decided whether to pass the actual value or a pointer
to it. Unlike C++, taking an argument by value doesn't imply a potentially
expensive copy, because arguments in Zig are immutable.

Note that sometimes taking an argument by value has semantic differences from
taking a const pointer, notably if the object is modified during execution.