Optimize Day 16 slightly by only entering the grid at places we did not exit from.
This is correct because if we enter at point p and exit at point q, then
the set of tiles illuminated by entering from q is necessarily a subset of
the set of tiles illuminated by entering from p.

This refactoring also allows removing all the integer casts, which is nice!